+++ src/engine/monitor.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/Chendemo12/functools/cronjob"
-	logger "github.com/Chendemo12/functools/zaplog"
 	"github.com/Chendemo12/micromq/src/proto"
 )
 
@@ -113,7 +112,7 @@ func (k *Monitor) closeRegisterTimeout(timeouts []TimeoutEvent) {
 	for _, c := range timeouts {
 		event := c
 		go func() {
-			logger.Info(fmt.Sprintf(
+			k.broker.Logger().Info(fmt.Sprintf(
 				"register timeout, actively close the connection with: %s", event.Addr,
 			))
 			k.broker.closeConnection(event.Addr) // 关闭过期连接
@@ -134,7 +133,7 @@ func (k *Monitor) closeHeartbeatTimeout(timeouts []TimeoutEvent) {
 	for _, c := range timeouts {
 		event := c
 		go func() {
-			logger.Info(fmt.Sprintf(
+			k.broker.Logger().Info(fmt.Sprintf(
 				"%s heartbeat timeout, actively close the connection with: %s",
 				event.LinkType, event.Addr,
 			))