+++ src/mq/mq.go
@@ -5,14 +5,11 @@ import (
 	"os"
 
 	"github.com/Chendemo12/fastapi"
-	"github.com/Chendemo12/fastapi/middleware/fiberWrapper"
-	"github.com/Chendemo12/fastapi/utils"
+	"github.com/Chendemo12/functools/logger"
 	"github.com/Chendemo12/functools/python"
-	logger "github.com/Chendemo12/functools/zaplog"
 	"github.com/Chendemo12/micromq/src/engine"
 	"github.com/Chendemo12/micromq/src/proto"
 	"github.com/Chendemo12/micromq/src/transfer"
-	"github.com/gofiber/fiber/v2"
 )
 
 var mq *MQ
@@ -24,11 +21,13 @@ type MQ struct {
 	broker   *engine.Engine
 	transfer transfer.Transfer
 	faster   *fastapi.FastApi
+	logger   logger.Iface
 }
 
 func (m *MQ) initBroker() *MQ {
 	// broker 各种事件触发器
 	m.conf.Broker.EventHandler = &CoreEventHandler{}
+	m.conf.Broker.Logger = m.Logger()
 	m.conf.Broker.Ctx = m.ctx
 
 	m.transfer = &transfer.TCPTransfer{}
@@ -39,47 +38,20 @@ func (m *MQ) initBroker() *MQ {
 	return m
 }
 
-func (m *MQ) initHttp() *MQ {
-	app := fastapi.New(fastapi.Config{
-		Title:                 m.conf.AppName,
-		Version:               m.conf.Version,
-		Description:           m.conf.AppName + " Api Service",
-		DisableSwagAutoCreate: !python.Any(!m.conf.StatisticDisabled, m.conf.Debug),
-	})
-	eng := fiberWrapper.Default(fiber.Config{
-		Prefork:       false,               // 多进程模式
-		CaseSensitive: true,                // 区分路由大小写
-		StrictRouting: true,                // 严格路由
-		ServerHeader:  m.conf.AppName,      // 服务器头
-		AppName:       m.conf.AppName,      // 设置为 Response.Header.Server 属性
-		ColorScheme:   fiber.DefaultColors, // 彩色输出
-		JSONEncoder:   utils.JsonMarshal,   // json序列化器
-		JSONDecoder:   utils.JsonUnmarshal, // json解码器
-		BodyLimit:     100 * 1024 * 1024,   // 设置请求体最大为 100MB
-	})
-	eng.App().Use(fiberWrapper.DefaultCORS)
-
-	app.SetMux(eng)
-	if python.Any(m.conf.EdgeEnabled, m.conf.Debug) {
-		app.IncludeRouter(&EdgeRouter{})
-	}
-
-	if python.Any(!m.conf.StatisticDisabled, m.conf.Debug) {
-		app.IncludeRouter(&StatRouter{})
-	}
-
-	m.faster = app
-
-	return m
-}
-
 func (m *MQ) Config() any { return m.conf }
+
+func (m *MQ) Logger() logger.Iface { return m.logger }
 
 // Ctx 获取根context
 func (m *MQ) Ctx() context.Context { return m.ctx }
 
 // Stat 获取统计信息类
 func (m *MQ) Stat() *engine.Statistic { return m.broker.Stat() }
+
+func (m *MQ) SetLogger(logger logger.Iface) *MQ {
+	m.logger = logger
+	return m
+}
 
 // SetCrypto 设置加解密器
 func (m *MQ) SetCrypto(crypto proto.Crypto) *MQ {
@@ -97,6 +69,10 @@ func (m *MQ) SetCryptoPlan(option string, key ...string) *MQ {
 
 // Serve 阻塞启动
 func (m *MQ) Serve() {
+	if m.logger == nil {
+		m.logger = logger.NewDefaultLogger()
+	}
+
 	// 初始化服务
 	m.initBroker()
 	m.broker.SetCrypto(m.conf.crypto)
@@ -107,12 +83,12 @@ func (m *MQ) Serve() {
 	go func() {
 		err := m.broker.Serve()
 		if err != nil {
-			logger.Error("broker starts failed: ", err)
+			m.Logger().Error("broker starts failed: ", err)
 			os.Exit(1)
 		}
 	}()
 
-	m.initHttp()
+	m.faster = CreateEdge(m.conf)
 	m.faster.Run(m.conf.EdgeHttpHost, m.conf.EdgeHttpPort)
 }
 
@@ -120,6 +96,40 @@ func (m *MQ) Stop() {
 	m.cancel()
 	m.broker.Stop()
 	m.faster.Shutdown()
+}
+
+// AddExchange 添加一个topic间的数据交换
+func (m *MQ) AddExchange(from, to string) error {
+	if from == to {
+		return engine.ErrExchangeSame
+	}
+
+	topic, exist := m.broker.FindTopic(from)
+	if !exist {
+		return engine.ErrSrcNotExist
+	}
+	dstTopic, exist := m.broker.FindTopic(to)
+	if !exist {
+		return engine.ErrDstNotExist
+	}
+
+	_, err := topic.AddForwarding(dstTopic)
+	return err
+}
+
+// DelExchange 删除一个topic间的数据交换
+func (m *MQ) DelExchange(from, to string) error {
+	if from == to {
+		return engine.ErrExchangeSame
+	}
+
+	topic, exist := m.broker.FindTopic(from)
+	if !exist {
+		return engine.ErrSrcNotExist
+	}
+
+	topic.DelForwarding([]byte(to))
+	return nil
 }
 
 func New(cs ...Config) *MQ {